Quickstart script cleanup

+++ deploy/custom-metrics-api/cleanup.sh
@@ -2,6 +2,11 @@ set -euo pipefail
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 cd $DIR
 
+if [ ! -f cm-adapter-serving-certs.yaml ];
+    then
+        exit 0
+fi
+
 rm apiserver.csr
 rm apiserver.pem
 rm apiserver-key.pem
